refactor(app): extract the export and split-pane wiring
Copying a detail field or the whole grid, CSV/JSON/SQL export,
cancelling a running stream or query, and the right pane's own grid
interactions move to `wire::split_pane`. The right pane mirrors the left
pane's filter/sort/copy/export behavior, so it reads better next to
itself than interleaved with the left pane's handlers.
