+++ app/src/wire/mod.rs
@@ -9,5 +9,6 @@ pub(crate) mod conn_form;
 pub(crate) mod edit;
 pub(crate) mod grid;
 pub(crate) mod settings;
+pub(crate) mod split_pane;
 pub(crate) mod tabs;
 pub(crate) mod update;
